Added copyright information at the time of the test files.
