--- src/sst/elements/vanadis/tests/basic-io/hello-world.c
@@ -1,6 +1,0 @@
-#include <stdio.h>
-
-int main( int argc, char* argv[] ) {
-	printf("Hello World from Vanadis\n");
-	return 0;
-}
