+++ src/sst/elements/vanadis/tests/small/basic-io/hello-world.c
@@ -1,0 +1,21 @@
+// Copyright 2009-2021 NTESS. Under the terms
+// of Contract DE-NA0003525 with NTESS, the U.S.
+// Government retains certain rights in this software.
+//
+// Copyright (c) 2009-2021, NTESS
+// All rights reserved.
+//
+// Portions are copyright of other developers:
+// See the file CONTRIBUTORS.TXT in the top level directory
+// the distribution for more information.
+//
+// This file is part of the SST software package. For license
+// information, see the LICENSE file in the top level directory of the
+// distribution.
+
+#include <stdio.h>
+
+int main( int argc, char* argv[] ) {
+	printf("Hello World from Vanadis\n");
+	return 0;
+}
